+++ src/providers/tripUpdates/lib/GtfsRTTripUpdateFeedProcessor.ts
